add variable prevUrlParamsArr and new code for updating urlParams
Checking the urlParams against a previous version of itself basically lets us know that there was a back OR fwd button press because these seem to be the only situations when it is not updated in an already existing method. So when the current and previous urlParams are identical, this means back or fwd was pressed, in which case this is the only time it is updated with the values extracted from the historically logged href string.